iss_energyObservable: corrected miscalculation of intra-vs-inter-molecular energies. Optimized random seed input: type long

+++ random.cpp
@@ -7,9 +7,9 @@ std::mt19937 my_rng {};     // defines an engine
 
 // Function to seed the random number generator from main file
 // Negative value uses TIME as random seed
-long unsigned int Seed(int seed) {
+long int Seed(long int seed) {
     if (seed < 0) {
-        long rseed=static_cast<long unsigned int>(std::chrono::high_resolution_clock::now().time_since_epoch().count());
+        long rseed=static_cast<long int>(std::chrono::high_resolution_clock::now().time_since_epoch().count());
         //std::cerr << "Randomizing random generator, seed is "<<rseed<<std::endl;
         my_rng.seed(rseed - seed);
         return rseed - seed;
